added a way to set new storage path directory

+++ src/csStorage/csStorage.UnitTest/Shared/TestAfter.cs
@@ -2,7 +2,7 @@
 using System.Reflection;
 using Xunit.Sdk;
 
-namespace csStorage.IntegrationTest.Shared;
+namespace csStorage.UnitTest.Shared;
 
 [AttributeUsage(AttributeTargets.Class | AttributeTargets.Method, AllowMultiple = false, Inherited = true)]
 public class TestAfter : BeforeAfterTestAttribute
@@ -10,10 +10,12 @@ public class TestAfter : BeforeAfterTestAttribute
     public override void After(MethodInfo methodUnderTest)
     {
         var directoryPath = new csContextBuilder<UserEntityMock>().DirectoryPath;
+        var inheritedDirectoryPath = new InheritedCsContextBuilder<UserEntityMock>().DirectoryPath;
 
-        if (Directory.Exists(directoryPath))
+        if (Directory.Exists(directoryPath) && Directory.Exists(inheritedDirectoryPath))
         {
             Directory.Delete(directoryPath, true);
+            Directory.Delete(inheritedDirectoryPath, true);
         }
     }
 }
